quant_realtime_dashboard/requirements.txt
streamlit
websocket-client
pandas
numpy
plotly
statsmodels
scipy

